Update preface.adoc
I'm currently revising my guide on database access in Java and was thinking - with the risk of it being rejected for self promotion- that including it in the 'Getting Started' guide would help more people get a clearer picture of the whole Java database landscape and where and how Hibernate fits in. At least that's the feedback I got so far.